#7 Update regex parse of tag commit name
The previous regex was replaced by ^v\d+\.\d+\.\d+$. However, that
syntax is an Extended Regular Expression (which the inbuilt `=~`
comparator does not support; it is based on the POSIX regex standard
instead). `grep` with the `-E` flag is used to parse the extended regex
as described in the comment.