#12 Add `utils` module, add `DEBUG` mode
This commit introduces `DEBUG`, wherein with a single macro, the
behavior of the program can be toggled between the debug and the
production mode. In the debug mode, the program will display the
zones on the image stream, which can be used for debugging purposes.

This is an initial version of the idea, and it will be further
leveraged in the future commits. Furthermore, the previous `debug`
function is now extracted to `displayZones` function in the `utils`
module.