#12, #15, #16 Update codebase, refineme graph generation
This commit largely updates the `app.py` file. It adds `argparse`
to support a seamless command-line argument parsing. These arguments
are now supported:
- `--turns-only` (to count only turns or no) [default=None]
- `--graph` (compute a graph of the accuracy) [default=None]

Furthermore, the graph generation is extracted to a stand-alone
module, `graph-generator`. This module reads the CSV file generated
by the main application and plots the graph of the actual and
predicted steering angles. The reason why it is extracted is to
that previously it would exhaustively generate the graph on every
frame parsed, which is not necessary.