#1 Purge conflicting git history
To manually remove any trace of *.joblib files that exceed the file
size limit of GitHub. Please, see #1 for more information. This will
impact the git history. The actual git history will be preserved in the
internal GitLab repository. The public (this) repository will have a
trimmed history of the project.